requirements.txt
Flask>=2.3.0
pdfplumber>=0.11.0
sentence-transformers==2.2.2
faiss-cpu>=1.12.0
requests>=2.31.0
python-dotenv
numpy>=1.26.0
tqdm
huggingface-hub>=0.19.3,<1.0
tokenizers>=0.15.0
psutil>=5.9.0
transformers==4.37.2
tensorflow==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
protobuf>=4.25.3,<5.0.0
packaging>=20.9
